Add a Class to manage libcurl
Libcurl is now managed in a single class. It takes care of cleaning up after requests.
A new function has been introduced to first get the header of a stats file. It then
checks if the local stats file is older than the remote one and only in the case of this
downloads a fesh stats file.